refactor: separate dropdown menu content from button
