+++ src/features/groups/components/GroupCardList/GroupCard/GroupCardContent/GroupCardDropdownMenuTriggerButton/GroupCardDropDownMenuContent.tsx
@@ -1,0 +1,40 @@
+import { DeleteIcon, PenIcon } from '@/assets/images/icons';
+import {
+  DropdownMenuButton,
+  DropdownMenuButtonIcon,
+  DropdownMenuButtonText,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  Icon,
+} from '@/components/ui';
+
+import React, { FC } from 'react';
+
+interface IGroupCardDropdownMenuContentProps {
+  handleRenameClick: () => void;
+}
+
+export const GroupCardDropDownMenuContent: FC<IGroupCardDropdownMenuContentProps> = ({
+  handleRenameClick,
+}) => {
+  return (
+    <DropdownMenuContent>
+      <DropdownMenuItem asChild>
+        <DropdownMenuButton onClick={handleRenameClick}>
+          <DropdownMenuButtonIcon>
+            <Icon icon={PenIcon} size={5} color="primary" />
+          </DropdownMenuButtonIcon>
+          <DropdownMenuButtonText>Rename</DropdownMenuButtonText>
+        </DropdownMenuButton>
+      </DropdownMenuItem>
+      <DropdownMenuItem asChild>
+        <DropdownMenuButton onClick={() => {}}>
+          <DropdownMenuButtonIcon>
+            <Icon icon={DeleteIcon} size={5} color="danger" />
+          </DropdownMenuButtonIcon>
+          <DropdownMenuButtonText>Delete</DropdownMenuButtonText>
+        </DropdownMenuButton>
+      </DropdownMenuItem>
+    </DropdownMenuContent>
+  );
+};
